scripts/data-migration/requirements.txt
psycopg2-binary==2.9.10
pymongo==4.9.1
pydantic==2.9.0
python-dotenv==1.0.1
